Extracted server auth logic to a separate module
