+++ src/Server.hs
@@ -1,20 +1,16 @@
 module Server where
 
 import API.Server
-import AuthData
-import AuthenticatedUser
 import Control.Monad.Logger
 import Control.Monad.Reader
 import Crypto.JOSE
 import Data.Aeson
-import Data.Password
-import Database.Persist
 import Database.Persist.Sql
 import Log
 import Network.Wai.Handler.Warp
-import Servant hiding (Unique)
-import qualified Servant.Auth as SA
+import Servant
 import Servant.Auth.Server
+import Server.Auth
 import Server.Config
 import Server.DB
 import Server.Endpoints.AuthTest
@@ -24,34 +20,6 @@ import Server.Endpoints.PublicProfile
 import Server.Endpoints.Register
 import Server.Endpoints.UpdateProfile
 import Server.Types
-
-lookupUser ::
-  ConnectionPool ->
-  AuthData ->
-  IO (Maybe AuthenticatedUser)
-lookupUser connPool (AuthData username password) = flip runSqlPool connPool $ do
-  mEntity <- getBy $ UniqueLogin username
-  let mUser = entityVal <$> mEntity
-      hashFromDB = userPasswordHash <$> mUser
-  case checkPass (Pass password) <$> hashFromDB of
-    Just PassCheckSucc -> return $ AUser . userUuid <$> mUser
-    _ -> return Nothing
-
-basicAuthCheck :: ConnectionPool -> BasicAuthCheck AuthenticatedUser
-basicAuthCheck connPool = BasicAuthCheck $ \basicAuth ->
-  maybe Unauthorized Authorized <$> lookupUser connPool (basicToAuthData basicAuth)
-
-authCheck ::
-  ConnectionPool ->
-  BasicAuthData ->
-  IO (AuthResult AuthenticatedUser)
-authCheck connPool basicAuth =
-  maybe Indefinite Authenticated <$> lookupUser connPool (basicToAuthData basicAuth)
-
-type instance BasicAuthCfg = BasicAuthData -> IO (AuthResult AuthenticatedUser)
-
-instance FromBasicAuthData AuthenticatedUser where
-  fromBasicAuthData authData authCheckFunction = authCheckFunction authData
 
 server3 :: JWTSettings -> ServerT API HandlerT
 server3 jwtSetting =
